Fixes bootstrapping problems on IPv6 enabled hosts. Partially fixes bug #3376322.
When a host is both IPv4 and IPv6 enabled, and an DNS lookup performed on a IPv4+6 enabled UHC cache, the DNS would return
both the IPv4 and the IPv6 address. GTKG would choose one of these addresses on random. However if the IPv6 enabled client
has only IPv4 internet connectivity it would not be able to reach the IPv6 address. It was possible that GTKG would randomly
choose the IPv6 address over and over again. After 3 attempts the UHC would not be retried any more and bootstrapping became
impossible.