IPv6-Ready: added support for the new "IP" feature.
Also extended the node handshakes to propagate the GUID of the host
to be able to spot duplicate connections (since a host can listen
on both IPv4 and IPv6 and we don't want to connect twice to the same
node).

Download mesh alternate locations can now list IPv4 or IPv6 addresses.

Made sure we skip IPv6 addresses in the UDP crawler pong responses
since extensions to that message are not yet defined by the IPv6-Ready
specifications.