XML parser was not handling illegal byte sequences in UTF-8 buffers.
When vxml_read_char() failed to decode a character, it wrongly assumed
it was due to improper truncation of a file buffer but did not handle
the case of invalid character byte sequence in a memory buffer.  This
caused a crash because the code assumed it was dealing with file buffers
and was expecting another input in the buffer chain.

Added an explicit VXML_E_ILLEGAL_CHAR_BYTE_SEQUENCE error and let
vxml_read_char() raise a fatal error with that error code when reaching
an invalid byte sequence.