Added malloc() replacement.
The new xmalloc() allocator can be a drop-in replacement for libc's malloc().
Its advantage is that it is based on top of the VMM layer and it fights hard
to limit memory fragmentation. Its disadvantage is that freeing blocks is more
costly than using the system's free() because of the systematic coalescing
attempts performed over the freelist.

By default, xmalloc() is now used as the main malloc() which means glib, GTK+
and friends from the libc allocate memory there and no longer perform any
heap-based allocation, excepted at startup time until the VMM layer is up.

It will be easy to turn off this superseding by editing a single line in
the lib/xmalloc.c file, to suppress the definition of XMALLOC_IS_MALLOC.

For long-term running applications, relying on a coalescing xmalloc() should
avoid the heap fragmentation that inevitably develops. Naturally, this defers
some of the fragmentation problems to the VMM layer but the latter is already
designed to limit it as much as possible.