Hack for FreeBSD.
It seems that glib 2.x calls us to free() blocks that have never
been malloc()'ed before and seem to come from an mmap()ed zone.
That's weird and will need to be investigated.

The hack here allows running on FreeBSD with xmalloc() but is
sub-optimial as all these blocks that glib frees are actually leaking.