Fixed harmful cleanup of routing table when cycling over naturally.
When we reached back slot 0 in the table through natural increasing and it
turned out that the time to cycle over was overdue, we would free up the
whole table.  If we were allocating a new slot from revitalize_entry(), that
meant a crash would follow immediately due to segmentation violation (accessing
unmapped memory).

Also optimized the logic of get_next_slot() with branch likelyhood hints.