Re-fork a child if we get a harmful signal.
When debugging hooks are run in the child, they may encounter corrupted
data and cause a segmentation fault, a floating point exception or whatever
harmful condition causing an immediate stop.

In that case, the parent will attempt to re-fork a child, disabling debugging
hooks this time, so that we can get back a full gdb stacktrace.