Added polymorphic logging interface, through a log agent.
The log agent is a polymorphic dispatcher to allow transparent logging to
stderr or to a string.  This allows one to write a logging routine that can
be used to either write things to the log files or to generate a string
without any timestamp and logging level information.

File loggging through log agent uses a non-NULL logthread object.
The aim is to avoid any memory allocation during logging that could invoke
malloc(), in case we are logging within xmalloc() and the latter is actually
replacing the former.