Added VMM counter for successful high-order cached pages coalescing.
When looking for consecutive pages that form a range larger than what we
can cache, we rely on adjacent high-order regions present in the cache
to be coalesced together.  Count the amount of times this happens in the
lifetime of a session.

Also fixed some comments and enriched debugging statements in the
vpc_find_pages() routine.