Let xmalloc() use walloc() for small objects.
The xpmalloc() interface can be used to allocate "plain" memory and make
sure we do not recurse back to zalloc() or walloc(): it only allocates
blocks using the VMM layer or the existing blocks in the freelist.

The advantage of using walloc() is that it drastically reduces the
fragmentation level of the freelist and avoids running costly coalescing
checks for the majority of the small blocks we allocate.  Lower freelist
fragmentation also means less space wasted since we don't need to hold
large split blocks that will mostly never get used.