Made s_xxx() and t_xxx() standalone logging routines.
There is no longer any redirection to g_logv() when not recursing or
not in a situation where we want to avoid malloc().

We therefore perform our own message formatting, which guarantees that
our logging routines will always support our specific implementation
of %m error message formatting: we include the symbolic errno value
in the message, whereas glib's version only expands %m with the output
of strerror().