On Windows, propagate the PC at which the exception occurs in the crash log.
There may be an error during the stacktrace dump causing a loss of the
location where the exception occurred: either because we do not manage to
unwind the stack properly, or because there is a subsequent crash before we
get to that point.

Therefore, log the PC as soon as possible, knowing that it comes from a trusted
source: the exception record generated by the kernel.