vmf_to_string(): use str_bprintf() instead of gm_snprintf().
On Windows, the vsnprintf() routine formats "%p" differently than the way
we use in str_vncatf(): no leading "0x", and hexadecimal letters are all
upper-cased.

Since the log messages in the VMM layer use the s_xxx() logging routines,
which rely on str_vncatf() for formatting, it's best to switch to the
str_bprintf() routine to have a coherent set of formatting for the VMM logs.