crash_handler(): allow auto-restart of process as an alternative.
At the end of the crash road, if there is no chance to dump a core or if
auto-restart was explicitly requested, attempt to re-exec() the process to
keep the service up.

To avoid endless loops due to frequent quickly happening crashes,
auto-restart only happens when the process has been up for 10 minutes.

Auto-restarts are enabled at crash_init() time with CRASH_F_RESTART.