Added Windows startup debugging support and fixed Cygwin startup.
Defining MINGW_STARTUP_DEBUG in the lib/mingw32.c file will create a file
called gtkg-log.txt at the place where gtk-gnutella is launched and tracing
how the early init performs.

When AttachConsole() succeeds, the resetting of stdio must be performed
only when stdin is a console, otherwise the re-opening fails and we're
no longer running with any input/output abilities.  This is usually fine
when --log-stderr and --log-stdout are used, but we would be totally blind
until then, should a fatal error occurr.