Fixed handling of relative position lists, causing unfair scheduling.
The insertion in the list was wrongly done: the relative order must be
a consequence of the absolute position in the queue.

Also by construction new upload items are appended to the list and therefore
they fall at the tail of the relative-position list.

Due to this new (but proper) logic, we now have to recompute the absolute
positions each time we remove an entry from the queue.

Removed useless parq_upload_update_relative_position_after() thanks to
the fixed logic.