Separate hash_table_clear() and hash_table_reset() operations.
Clearing means removing all items stored in the table, resetting means
freeing up the memory used by the table to store its data.

The hash_table_clear() operation can be offered to clients.

The hash_table_reset() is an internal operation which does not need to call
hash_table_clear() first since we're disposing of the memory anyway so there
is no need to painfully reconstruct the bin free list!