Limit length of stackframe atoms to routines belonging to our own text.
By chopping off trailing parts of the frames that refer to shared-library
routines, we limit the size of the atoms we generate and also their amount.
Since these atoms are never freed once allocated, it saves a significant
amount of memory when tracking leaks whilst not limiting the ability to see
the significant part of the calling frame.