Removed harmful assertion causing startup crash.
When zalloc_memusage_init() is called, we iterate on an hash table that
can be modified through calls to memusage_alloc().  The behaviour is not
going to be the same on 32-bit and 64-bit systems due to different object
sizes being involved.

So rather than asserting that no zone can be setup already for memory
usage collection, check that when there is something then it is a valid
object.