Added xsort() routine avoiding any malloc() call.
There are times where libc's qsort() will use malloc() because it is
implemented as a merge sort instead of quick sort.  However, we need to
have a sorting routine capable of sorting without calling malloc().

This implementation is an extract of the qsort() implementation of the GNU
C library, adapted for inclusion in gtk-gnutella and tailored to use the VMM
layer instead of malloc().

This makes it safe to use in zalloc() and other low-level code where we may
not want to have malloc() invoked during a sort operation.