pmap_insert_region(): must extend pmap if it becomes full after insertion.
It was wrong to check for a full pmap upon entry in pmap_insert_region()
because pmap_extend() allocates a new memory region based on the current
pmap, which of course does not contain the region we are about to insert.

Because the allocation of the pmap happens without immediately updating
the pmap itself, it could possibly happen that the new pmap allocated
overlapped with the region we had to initially insert in the pmap.

Hence it is necessary to make sure pmap_insert_region() will always have
a free slot for inserting regions, calling pmap_extend() after processing
the insertion in case we filled the last slot.