Fixed GUESS corner-case crash.
When processing a GUESS RPC timeout and iterating on a starving query that
has been required to end as soon as it starves, the query was destroyed
too soon, the RPC callback not expecting to have the object released in the
middle of its processing, resulting in an assertion failure.

This fix introduced asynchronous termination.