Fixed quadratic beahviour of quicksort() when given identical items.
When all the items in the array are identical, the pivot is necessarily
equal to all the items in the partition and the old code was not prepared
for this situation, leading to O(n^2) running time.

The fix involves swapping the items identical to the pivot to the head/tail
of the partition and later moving back the items to the center of the partition
before handling the left/right sides.  Although this involves more swaps,
it keep the complexity closer to O(n.log n).