Added dynamic linker symbol probing support.
When the BFD library is not available, we can fallback on the dynamic linker
to extract the publicly-available symbols from a shared object.

The dynamic linker is also used to find out which loaded file can hold
a particular symbol so that we can in turn use the BFD library to extract
the symbol information.