Added the concept of internal key reference within a value structure.
For situations where the key is a reference field within a value, there is
no need to explicit the link between the key and that value: the structure
creates the implicit link.

This allows creation of (key, value) tuples indexed by keys by saving
one pointer per item compared to regular hash tables.