On Windows, be smarter when reserving the initial amount of virtual space.
We do not want to have too little "unreserved" VM space because that space
will be fragemented by Windows and we may end-up with not enough consecutive
space to allocate memory when no hint is supplied.

Log the amount of reserved memory and its range at startup, as well as the
amount we left available.