Added thread_is_stack_pointer() for safe backtracing from another thread.
The thread support runtime now tracks the stack base of each thread it knows
about, to allow one to determine which thread a stack pointer belongs to.

This is used by vmm_is_stack_pointer() when it detects we're running in a
multi-threaded environment in order to allow proper backtracing of a given
stack from another thread.