xmalloc: add "direct" locking of unshared thread-specific pools.
The aim is to signal that we're in a direct allocation path, so that a
concurrent free can be correctly handled, the concurrent thread accessing
the thread-specific pool taking the time to wait for the lock.

This does not completely close the race condition, but further limits the
window of opportunity for corruption.