posix_memalign(): allocate blocks between 1/2 and 1 page from a new VMM page.
Also relaxed the xmalloc_is_valid_length() condition to make it possible to
actually implement posix_memalign() correctly for large blocks.  The constraint
of having to always use a valid block size for the allocated block was too
great.

It is now considered valid to have a block size that is not exactly fitting
a bucket for sizes larger than XMALLOC_FACTOR_MAXSIZE.