stacktrace_chop_length(): don't chop stack unless stacktrace_init() was called.
Until they called stacktrace_init(), we don't know whether we're running
as part of a statically linked program or whether the whole program is
held in shared libraries, the main() entry point being just there to load
the initial shared libraray.

This means stack_is_our_text() is unsafe in its current form.

Note: this precaution is only there to allow the library to be used outside
of gtk-gnutella as a "service provider", where we cannot assume anything
about the layout of the program.