Adjusted initialization procedures for FreeBSD pthread idiosyncracies.
On FreeBSD, an initial pthread_self() call can call malloc(), and furthermore
the pthread package plays tricks with threads during the intiialization,
making the value of pthread_self() change in the main thread.

As a counter-measure, we now detect threads whose stacks either overlap with
existing threads or fall within another thread (supposedely deceased, but
that is an event we cannot see) to be able to reuse their thread element.