xmalloc: wait for the VMM layer to be initialized before installing xgc().
This is an attempt to tweak the initialization order of the library to see
whether recent problems reported on 64-bit platforms were linked to the early
auto-initialization of the callout queue.

Waiting for the VMM layer to be up before auto-installing the xgc() idle
callback further guarantees that we'll avoid possible sbrk()-based early
allocation.

Besides, there's no real urgency into installing the xgc() callback, so it is
perfectly fine to wait for the VMM layer to be up to do this.