Relax file_path_set() and open_read() a bit to gracefully handle relative paths.
Apparently bug #3534002 exhibits a problem with a pre-compiled binary on Debian
that does not supply absolute paths for some configuration files, which prevents
any startup.

Moreover, there is not enough data in the logs to identify who the culprit is,
so we are now loudly complaining when we're given a non-absolute path, without
necessarily crashing immediately.  We will crash later if the first path to
load a file from is not absolute, but otherwise we shall continue, with possibly
a degraded service (e.g. geo IP data missing, or no hostile support).