bsched_enough_up_bandwidth(): adjusted logic for non-sharing servents.
When a servent is not uploading anything (either because uploads are disabled
or because none of the shared files have been served in the last hour), there
is a potential for becoming an ultra node.

Furthermore, when no bandwidth limits are set for Gnutella and HTTP, assume it
is OK to become an ultra node.  The old code was skipping this test, hence it
refused to promote peers with enough bandwidth to become one.