Extended the ID table to specify the size of the generated IDs, in bits.
Added idtable_try_new_id() now that the ID table can be arbitrarily small:
the idtable_new_id() routine will abort execution when the table is full so
we need a way to report that situation, provided the client code is ready
to handle it.

Hence a new routine, unused in legacy code where the table has its maximum
size and we don't expect it to get anywhere near full..

Another important change is the forced rotation of the allocated IDs throughout
the whole ID space.  In other words, an ID is not reused until we have already
allocated all the other available IDs.