Added magic number to wrap_io_t and use it to check socket is still valid.
For code that is not seeing the gnutella_socket directly but only gets a
wrap_io_t address, we need a convenient way to make sure the socket is still
valid and has not been destroyed.  That code cannot know that wio->ctx is
the actual socket, hence the need for a magic number.