Plugged support for "OOB Proxy Veto" -- now ignores "NP" in queries.
The "NP" (no OOB proxying) was sent by legacy leaf servents to tell their
ultrapeer that they should not proxy their requests (in case they do not
support secure OOB).

This is now handled through the "OOB Proxy Veto" vendor message, which we
processed correctly (in that we recorded whether or not to OOB proxy queries
for a leaf) but did not otherwise honour.