parq_upload_load_queue(): was not properly handling empty tag names.
Since PARQ-capable entries were emitted with an empty "PARQ:" tag and the
parsing bug caused the remaining of the PARQ state to be ignored, we were
not properly restoring the persisted PARQ queue upon restart!

We are now explicitly catching empty values, and can therefore emit the PARQ
tag without any space after the ":" character.