acct_net_update(): don't fail if called with pmone == -1 and entry is missing.
Due to DHT node collisions, we can have two different IP addresses for a node
and learn the new one after it was already inserted in the lookup path.
If the new address was not in the same class C network, we could have an
assertion failure later.

This was a case of over-protection when the DHT code was written, asserting
something at a distance and then having the code evolve, silently breaking
this implicit (and now invalid) assertion.