Fixed wrong DHT value response message format.
This was a severe bug:

We forgot to send a trailing 0 in the value response to indicate that there are
no secondary keys, and GTKG was reading those responses as badly formatted and
therefore ignored them!

We are now more lenient in our parsing, treating the missing trailing 0 as an
indication that there are no secondary keys (thereby parsing messages from
legacy GTKGs correctly) and also emit the trailing 0 now (so that legacy GTKGs
can understand the message properly).