Make sure RUDP traffic is caught at a lower level and not accounted as UDP.
The RUDP layer is used to implement firewalled-to-firewalled transfers via
a mini TCP-like layer built on top of UDP.  Therefore, it is used as the
basis for higher-level connections (HTTP) and will have to be accounted
for once the type of traffic is known, by upper layers.

We nonetheless track in the statistics all the received (and sent, at a later
time) traffic that pertains to the RUDP layer.