Do not route duplicate queries with higher TTL to leaves.
We need to route duplicate queries with higher TTL to our neighbouring ultra
nodes so that the reach of the query be corrected.  But our leaves already
got the query (if their QRP table allowed it) the first time, so there is no
need to resend it a second time -- leaves do not forward queries.