Prevent sending of duplicate OOB-proxied queries.
We forgot to record the TTL of the message when issuing an OOB-proxied query.
Later, when receiving duplicate echoes from other nodes, we thought these
were sent with a higher TTL, and we re-forwarded them (because we were
comparing the message's TTL with 0).

This is a long-lasting bug (several years) that caused GTKG to spit out
duplicates needlessly...  Of course, neighbours trapped them, but this used
up some outgoing bandwidth!