Handle hostile addresses during DHT message reception for RPC timeouting.
A node could have its address become hostile dynamically whilst we have
an RPC in transit to that host.  We don't want to blindly drop the incoming
reply at the lowest UDP reception level.  Instead, we handle hostile checks
in kmsg_received() to be able to timeout pending RPCs if we have to drop
a reply.