Restored working auto-restart on Windows.
The problem was that using the native execve() caused mis-parsing of the
argv[0] argument when it has embedded spaces, like for instance:

	"C:\Program Files (x86)\gtk-gnutella\gtk-gnutella.exe"

The solution is to trap execve() on Windows to be given a chance to surround
the argv[0] within quotes, in order to let the C runtime parse the argument
correctly and not split it at space boundaries!

Of course, we should really handle ALL the arguments like that, but this will
have to wait another release.  For now the urgency was to restore the auto-
restart on crash before getting 0.98.4 out of the door.