Optimize fileinfo source switching during download_clone().
The old code was calling file_info_add_source() at download_clone() time and
then later the parent download was removed and we called the
file_info_remove_source() routine on it.

Meanwhile, the fileinfo was marked dirty, and we also caused a recompute of
the available ranges for the file twice (triggered by both source addition
and removal now), for nothing since there were no changes in the available
chunks.

So file_info_clone_source() was introduced to perform just the amount of work
required to switch the source from the original to the clone, and the original
download is flagged as having been cloned.  This will prevent a subsequent
file_info_remove_source() from updating the list of seen chunks on the network.