[GTK] Download source details were incorrect sometimes for queued downloads.
Due to a long-standing editing bug in download_gui_rescheduling(), all
sources for which the rescheduling starting point was in the past were not
correctly updated.  Instead, the buffered status from a previously displayed
source was shown.

With this bug fixed, all sources which have gone past their rescheduling
timestamp will be shown as "restartable since ...", whilst they are tagged
with a "rescheduled for ..." when they are before the rescheduling timestamp.