SDBM: delipair() now proceeeds with stripping even if big data cannot be freed.
The delipair() routine was wrongly aborting with an error condition as soon
as we failed to properly free big key/value information.  This caused the key
and its value to stay on the page, wrongly consuming space and also possibly
in a stage where the big key was cleaned up but not its data.

Therefore, we now always perform deletion on the page, regardless of whether
we could clean up big key/value information.