After unclean shutdown, rename existing core file in crash directory.
To that effect, we persist the PID of the current process to be able to
rename the core file appropriately with that PID as suffix.

This will prevent gtk-gnutella from overwriting a previous core file after
an auto-restart, which would severely limit the debugging of the original
problem when a subsequent crash occurs.