VMM: closed race condition during allocations.
There was a race condition during virtual memory allocation whereby we computed
a hold in the VM space and then used it as a hint to allocate memory.

This was fine in a mono-threaded application but this cannot be allowed in
a multi-threaded one.  The race was causing two problems:

- It forced vmm_mmap_anonymous() to perform an additional pmap_lookup() when
  the selected hint was not allocated (in case a concurrent thread sneaked in
  and already filled the hole with a successful allocation).

- It did not fit our simple virtual memory manager on Windows whereby the
  supplied hint is expected to be available.

The downside of fixing that race is that all virtual memory allocations are
now funnelled.  Fortunately, this does not prevent concurrent memory freeing
and when there is a page cache hit (most likely scenario), the critical section
does not span across a system call to get more memory.