Fixed sending duplicate message with higher TTL.
In some cases a duplicate message would be forwarded. This
would happen if the TTL+HOPS > max_ttl. Initially when checking
wether a message should be forwarded, only the TTL from the message
would be read. And if the new TTL was higher compared to the previous
know message. The message was marked to be forwarded.

When forwarding the message the TTL+HOPS > max_ttl check was done.
If the TTL+Hops was greated then max_ttl the TTL would be trimmed down.
However that could cause the TTL to be the same again as a previously
forwarded message.
Therefor, also check the max_ttl when determining wether the message
should be forwarded.