Added crude protection when writing important files on filesystems like "ext4".
Since "ext4" is using a deferred block-allocation strategy, we need to be
careful when writing new files that are renamed later, superseding an older
file.  This ensures that at least one of the versions (old or new) is still
available should there be a system crash before the data can be fully allocated
and flushed to disk.