PARQ: accept to move an upload to the "dead" queue even if we wait for QUEUE.
An upload could be flagged as waiting for a QUEUE message, yet there are so
many waiting uploads from that host that we're not contacting it back quickly
enough and all these slots remain "alive" in the queue, preventing real uploads
from starting.

To keep some fairness in the system, we need to expire these uplaods, keeping
them in the "dead" list for a while.   They won't be totally lost until we
run out of PARQ slots, but they will allow newer uploads to start.