Windows: added workaround for bug in the MinGW C startup code.
It appears dirname(), which is called by the C startup, is invoking free()
on a pointer that has not been allocated by our malloc(), hence causing a
fatal error because it triggers an assertion failure that cannot be processed
so early in the program startup.

The workaround consists in disabling free() until we know we're out of the
MinGW startup code, as soon as we enter mingw_early_init().