mingw_win2posix(): allocate the "warned" hashed-set as late as possible.
Also close a race condition during the warned set creation in multi-threaded
environments.

Deferring the creation of the "warned" set allows proper initialization on
Windows because the normal startup code does not require the set to be created.
Otherwise, there was an initialization loop because the routine was invoked
through the steps required for the hash_offset_init() routine, due to an
error (expected) in fstat() during entropy collection.