xmalloc: guard bucket extension with a flag to prevent large recursions.
When now explicitly flag when we expand a bucket, and avoid attempts to
insert more blocks into an expanding bucket: now that we can defer frees
to a bucket, we have a safe way out of the recursion.

This should help keeping the thread stack usage down to a reasonable level.

The same reasoning applies to the processing of deferred blocks in a bucket:
since we have many places that can detect and trigger retrofiting of these
blocks back to the bucket, we indicate that we are in the process of handling
these blocks to avoid needless recursions.