Prevent deadlocks in rand31() by allowing lock-free seeding checks.
A deadlock could occur in the previous code if rand31() was called very
early in the process, due to the auto-initialization sequence that creates
a time thread and initializes the hash offset using... entropy and rand31().

Also renamed THREAD_ as RAND31_ to prevent ctags conflicts and be consistent
with what is done in other files.