Reduced latency when waiting for spinlocks or rwlocks.
We used to busy-loop for a while, then sleep for 2 ms whilst waiting for the
lock.  The latter has been reduced by 10, to 200 us.

This dramatically reduces the overall execution time when having heavy
contention on the locks and the lock is kept for a reasonably small amount
of time, which should mostly be the case in practice.