Renamed master waiter boolean fields to avoid confusion with waiter's field.
In particular, both structures had a "notified" field and this could lead to
obfuscated code when dealing with both the waiter and the master waiter in
the same routine.

For consistency, the master waiter boolean fields are now prefixed with "m_"
to indicate they refer to the master part of the structure.