xmalloc: improved aligned zone allocator to use fine-grained locks.
The locking strategy was wrong. Some lookups in the aligned[] array were
done without locks if we were not able to grab the spinlock.  The old code
did this lazy spinlock grabbing because of possible recursion, but we switched
to a mutex to make sure we always lock the aligned[] array.

The "zone allocator" is the internal simple allocator used to allocate aligned
memory which is "small-enough".  It was redesigned to clearly add a wrapping
xzone structure, leaving the xdesc_zone descriptors for the allocation arenas
within the zone.

This allows finer locks since we can now lock only the zone to which we
allocate, leaving ghe global xmalloc_zone_slk spinlock to protect the xzones[]
array.

The hand-written linking of xdesc_zone descriptors has also been replaced
with an embedded list, to factorize code and make it easier to read.

Since the xzones[] array is never freed, nor are the xzone structures, we now
allocate them via omalloc().