Make TTH recomputation of shared file prioritary if it does not have a TTH yet.
We call request_tigertree() on all the shared files to defer to the background
verification task the cost of probing the cached TTH entries and determining
whether they are up-to-date.  This means after a library rescan the queue is
filling up just to ensure our cache is OK.

However, sometimes we add a new file for sharing, or shared files may not have
their TTH known yet. The TTH computation for these files should be made a
priority, i.e. their request should be put ahead of the queue.