Added s_xxx_once_per() logging routines to limit emission to given period.
In case an I/O error occurs on the SDBM back-end, we do not want to get a
flood of error messages for each access. Limit logging to a reasonable period.

Also protected values_publish() and values_get() against sudden I/O errors
in the DHT raw database, to safely catch the condition and return an error.