filename_sanitize(): replace "()" with "[]" in filenames for easier cut-n-paste.
Parentheses are a problem for most UNIX shells since they introduce escape
sequences like $() or indicate that a sub-shell should be used, or that a
function is being defined.

Therefore, to ease copy-n-paste of filenames without going through the shell
auto-completion, which is not always convenient, it is best to avoid parentheses
altogether.