download_mark_active(): do not reset fi->recv_last_rate on HTTP follow-ups.
For regular download where we have only a single source and we face a
follow-up HTTP request, we must not reset fi->recv_last_rate or we'll
mess up with the average download speed estimate for the file.

This fixed a bug in the GUI where the estimated completion time was reset
each time we were issuing a follow-up request for single-sourced downloads.