Added thread_wait() and friends to wait for thread termination.
These calls allow the calling thread to block until the specified thread
has exited, even if it is a detached thread (with some caveat since there can
be a race condition).

Using thread_wait() on detached threads is useful at process shutdown time when
we know we will no longer create threads and we still want an easy way to
synchronize with these detached threads.