Allow file verification (SHA1 and TTH computations) to be done in other threads.
To speed-up file hashing, processing is now be performed in dedicated threads,
transparently for the client thread (which does not need any particular lock
protection to handle callbacks from the verification thread, as if processing
continued to be done by a local background task).

Also fixed verify_enqueue() which was wrongly returning that an item had been
inserted if it was a duplicate with low priority, causing extra references to
be kept on the shared files, and therefore leaking memory.