Switched halloc() back to walloc().
We had switched halloc() to xpmalloc() because, at the time, walloc() had an
incorrect implementation of thread-specific zones not supporting freeing from
a foreign thread.  Then the thread-specific zone support was dropped because
it was incorrect and halloc() remained linked to xpmalloc().

Now that walloc() uses thread magazines to distribute its objects, it is
perfectly efficient in a multi-threaded environment and it makes sense to
base halloc() on walloc() again.