zalloc: added zone_size() to return user (logical) block size for a zone.
When compiling with ZONE_SAFE or other TRACK_ZALLOC, which add a header to
all the allocated blocks, the user pointer is not at the start of the physical
block and we need to distinguish between the actual block sizes in the zone and
the usable portion of each block for user data.

The old zone_blocksize() routine returns the physical block size, whereas the
new zone_size() returns the logical block size.  The two are identical usually,
when there is no tracking / debugging overhead added to blocks.

Also fixed an assertion in zfree() when compiling with ZALLOC_SAFETY_ASSERT
and ZONE_SAFE at the same time.