thread_find_tid(): use thread_allocated_stid instead of thread_next_stid.
The thread_allocated_stid variable is atomically updated each time we are
increasing the maximum thread ID, whereas thread_next_stid is a computed
variable, after the creation was done.  To allow thread_find_tid() to work
correctly for a launched thread, we need to rely on thread_allocated_stid.