Strengthen entropy_rand31() by initializing a non-zero "previous entropy" array.
The entropy_rand31() routine is used to randomly shuffle the various arrays
that let us combine the collected information.  It is based on the weak rand31()
generator, but combines the output with previously collected entropy.

We were not initializing this previous entropy variable, hence it was zero,
meaning we were not benefiting from the stronger randomness for shuffling in
the initial collection stage.

We now therefore initialize the array to a "random" value, and strengthen the
alteration of rand31() output by constantly shuffling our previously collected
entropy.