xmalloc: defer call to xpages_pool_init() until we need it.
This is to avoid a deadly init recursion on FreeBSD for evq_init().

We had to refactor once_flag_run() a little bit: it now returns nothing.

The new once_flag_run_safe() routine can be used when there is a risk of
recursive initialization and it will return FALSE when it detects such
recursion.

For symetry, we added once_flag_runwait_safe() as well although we don't
use it for now.