Enhanced the file_object caching layer to deal with read-only files.
Internally, we want to be able to open all our files in O_RDWR mode, to later
serve users wanting to read, write, or do both on the file.  But if the file
on the filesystem is read-only for instance (but the same goes for write-only
files), we cannot internally open the file in O_RDWR mode).

In that case, we can only open the file internally to let users do the I/Os
they want, provided the filesystem permissions allow this kind of access of
course.