Made the QRP background task management thread-safe.
Thread-safety was necessary because the qrp_prepare_computation() routine
is called from the library thread, and we do not want to use a TEQ event
to schedule it.

As a consequence, this will also shut up the warning about "lost exit code"
for QRP background tasks since we now have a "done" handler.