forget_merged_node(): was incorrectly resetting status to KNODE_UNKNOWN.
This lead to assertion failures later in free_node_lists() when the old
(now merged) buckets were disposed of, because the nodes that were not kept
failed to bear the proper status in their respective list, an error caught
by check_leaf_list_consistency().

For security, added checks about the list consistency in merged_node_list()
to verify that the two lists we are merging are already holding nodes with
the correct status.