TEQ: must use new cq_zero_if_triggered() after setting throttling to close race.
There was a race condition when setting throttling on a TEQ because the event
could have been already dispatched upon return from evq_raw_insert() and we had
no way to check for that, leaving the queue in "throttled mode" because the
event was never cleared out.

The new cq_zero_if_triggered() call atomically checks whether an event has
already been triggered after insertion in the queue.