gnet_stats_count_received_payload(): must count header explicitly for G2.
Since G2 packets do not have a header as Gnutella ones, the routine
gnet_stats_count_received_header() is not called.  However, we need to
update some packet statistics, which is what is normally done in that
routine.

Therefore, split gnet_stats_count_received_header() to allow us to perform
the necessary accounting each time we receive a G2 packet.