Cleaned up Gnutella handshaking to emit an Accept if the remote sent one.
The negotiation was not complete at the handshaking level.  Normally, Gnutella
nodes don't bother with Accept, but if the remote end sends one, we should
also reply with one and emit the Content-Type header acknowledging which of
the acceptable entries we picked (that we did for G2, but not for Gnutella,
when an incoming connection says it supports both and we pick Gnutella).