gnet_stats_g2_count_flowc(): do not set TTL=0 when synthetizing a TTL for G2.
Since this is an outgoing path, no message is expected to be sent out with
a TTL of 0 in the message queue.  Because G2 messages do not have a TTL, we
have to synthetize one before calling gnet_stats_flowc_internal(), and of
course we must avoid setting it to 0, as we were before.