Do not attempt to OOB-proxy queries in search_request()!
This was old code that had been kept there and was a really bad idea: it was
messing up, in-place, with the MUID of the query and had unwarranted effects
when the dynamic query was launched afterwards.

Also, some leaves seem to perform dynamic querying of their own, sending
duplicate queries with higher TTL.  These need to be ignored and not passed
to the dynamic query layer.