Prefer entropy_minirand() to entropy_rand31() since the former behaves better.
The flaw can be witnessed by running:

	random-test -B -p 100000 -F 0xffffffff -1
	random-test -B -p 100000 -F 0xffffffff -2

The output of the former command exhibits an anomaly in bit #10, which does
not appear with the second command.  The flaw stems from the weakness of the
underlying rand31() algorithm, of course, but the composition done in the
entropy_minirand(), which is the same as the one done for rand31_u32(), seems
to yield better distribution. The proof of that origin comes from running:

	random-test -B -p 100000 -F 0xffffffff
	random-test -B -p 100000 -F 0xffffffff -u

which compares rand31() and rand31_u32() and leads to a similar flaw / fix.