Double the maximum size of array we can shuffle during entropy collection.
Because the number of bits used for the context of entropy_rand31() is 191,
we have largely enough entropy to reach all the permutations of arrays of
size 24, since 24! is about 2**79.04.

Being able to permute a larger size means we can shuffle more entropy gathering
routines at the same time, increasing the set of possible initial values.