Ditched entropy_rand31(), and made entropy_minirand() use the KISS PRNG.
Since the introduction of entropy_seed(), we were gathering extra bits of
random context, so it was counter-productive to keep using rand31() when we
are in a position to use a better PRNG, which only requires 128 bits of initial
context to run.

As a bonus, this lets us increase the maximum size for a shuffled array to 33,
since the period of the KISS algorithm is at least 2**125.