Added ALLOW_UNINIT_VALUES symbol to avoid using uninitialized bytes.
The uninitialized bytes are normally used to collect additional entropy from
the system, mostly to initialize the random seeds for pseudo-random number
generation.

Although this is perfectly safe (yes, the result is undefined, but that is the
point), it creates a cascade of errors when running under "valgrind" because
the random seeds are all tainted by the use of these undefined values, and
many operations in gtk-gnutella depend on them (e.g. the initial random hash
offsetting).

Therefore, when running under "valgrind", undefine the ALLOW_UNINIT_VALUES
symbol in common.h, to prevent these "errors" from being reported.