socket_plain_sendto(): when debugging, emit warning if we can't convert IP addr.
When sending to an IPv6 socket, we must be given an address convertible to IPv6,
and when sending to IPv4, we must have a (possibly convertible) IPv4 address as
the destination.  Otherwise, sendto() is going to fail.

Therefore, when debugging, make sure we emit a warning with a stacktrace to
identify the calling chain which would not give proper destination addresses
when attempting to send a UDP message.