Added support for queuing locks, with strict FIFO thread ordering.
These locks can be configured as either plain locks (aka spinlocks) or
recursive locks (aka mutexes) and guarantee that the waiting threads will
get access to the lock in a strict FIFO ordering, based on their arrival
on the lock.

It is also possible for a thread holding a queuing lock to rotate it among
the waiting threads, in effect releasing it and then re-acquiring it, albeit
in an optimized way, with little overhead when there is noone else waiting
for the lock.

thread_timed_block_self(): relaxed "no locks held" precondition.

Since qlocks are using that routine to sleep, and we want to allow qlocks() to
be grabbed with other locks held, possibly, we need to relax the precondition
and allow locks to be held if there is a timeout given and it is less than
the hardwired THREAD_SUSPEND_TIMEOUT seconds timeout.

This new precondition will be met by qlocks, but to make sure it is enforced
by static assertions, we need to make THREAD_SUSPEND_TIMEOUT globally visible.

Increased main thread blocking timeout to 20 seconds, to support qlocks.

Since qlocks use thread_timed_block_self() in their main waiting loop, we
need to ensure the main thread can block for at least as long as the qlock
main waiting time, which is 20 seconds.

thread_element_block_until(): be aware that we may now block with locks held.

With the introduction of qlocks, it is now possible to enter the thread
blocking code with locks already held by the thread.  Therefore, we must
not blindly dispatch signals but ensure there are no locks held already.