Added thread_deadlock_check() to perform deadlock detection.
When low-level locks such as spinlocks or mutexes wait for some time, instead
of issuing a loud "possible deadlock" warning, we are now invoking that new
routine to see whether we are in a deadlock situation.

If not in a deadlock, nothing happens and the waiting continues.  If it times
out later, then we will abort with an actual deadlock situation that we could
not detect (due to the presence of hidden locks or other blocking that we
do not consider such as condition variables, signals not coming, etc...).

If there is a deadlock detected, then we abort immediately, dumping the lock
stack of all the threads involved in the deadlock.