routing_chunk_move(): must check messages carefully when chunk was moved.
When a chunk is being re-allocated by the VMM layer (moved to a better place),
we go into each messages and update some references made to the chunk to which
the messages are currently stored.

However, during the validation of the slot pointer done before updating it, we
were checking the validity against the new chunk location, instead of the old
one.  The new message_check_chunk() routine is therefore given the old chunk
location so that we can validate the slot properly.
new chunk address, causing assertion failures.