thread_element_matches(): pay more attention to discovered threads.
When threads are created outside our control, many things can happen and we
only witness the consequences.  To be safe, we must refuse a QID cache match
for dicovered threads (other than the main thread) when we detect that the
internal thread ID has changed since the time we registered the thread.

This fixes a bug in Windows (at least, this is where it was easy to reproduce)
when trying to configure the download paths from the preference window: GTK
seems to be creating threads to handle the pop-up window for the path selection
and that was triggering the issue.