cond_wait_until(): now correctly handle underlying cond variable mutation.
When a waiter is attached to a condition variable, it is mutated into an
extended form capable of holding extra information.  Since this creates a new
object, cond_wait_until() must understand that a change in the condition
variable object is allowed when a mutation occurs, provided the same mutex
is used in the old and new object.

When the mutation is detected, the cleanup variable is updated so that we can
properly release the variable.