dir_entry_filename(): made the Windows version thread-safe.
Now that the routine uses h_private(), memory is lazily freed therefore we
removed the ability to call the routine with a NULL pointer to free memory.

On UNIX, the dir_entry_filename() was a macro and calling it with NULL had
no effect, causing Coverity to complain.  This fixes CID 106351 and 106352.