Added TX layer write-fault mechansim to help detect user error handling bugs.
Following commit e2f31af2fd63e1503117a59200a725885d2cbdf3 which was due
to improper error handling in the user code, the TX layer now has a write-fault
installed on upper TX layer when a write error happens in some lower layer.

The next write issued to the flagged TX layer will immediately be caught and
cause a loud warning with a stack trace, likely pointing to the culprit.