Windows: monitor committed memory, crashing before the kernel would kill us.
Unfortunately, on Windows, it is quite possible to run out of memory without
being able to notice it beforehand, since other parts of the process may
allocate memory without our knowledge.  And when that happens, there is a
nasty popup made that blocks the application until the user closes it.

That does not allow for clean (unattended) restarts. Therefore, we monitor
the amount of memory committed and force a crash when we reach dangerous
levels (about 90% of the unreserved memory that the process can use).