Adjusted lock tracking when locks are disabled, fixing lock swapping.
When thread locks are disabled, we are much more lenient with panics due
to resource exhaustion or not find locks, but nonetheless we track things
properly when we can, if only to be able to spot bugs in the error handling
paths.

When locks were disabled, we were not recording lock swaps at all, which would
cause warning messages on Windows during the crashing path, when the semaphore
array was cleaned-up, since that path uses lock swaps.