UPnP: protect upnp_mapping objects given to callbacks with reference counts.
When changing the listening port dynamically, exsting UPnP mappings are
revoked and new ones installed if needed.  However, we were freeing the
old mapping objects whilst the UPnP RPC was in transit, resulting in a crash
when the reply came back since the mapping object given as argument was gone.