Allow ftw_foreach() to gracefully handle ENFILE and EMFILE error conditions.
When we try to open a directory and we get either an ENFILE or EMFILE error,
we now attempt to close one opened descriptor among all the parents we are
currently traversing before retrying the corresponding opendir() operation.

Indeed, a shortage of file descriptors can happen at any time, regardless of
the amount of descriptors they have requested that ftw_foreach() use.