Use hashing_mix32() each time we compose simpler hashing functions.
Each time we compose hashed values and one of the intermediate values comes
from a function that is not performing bit avalanching, we use an extra
hashing_mix32() step on the final result to increase dispersion.

This is true for pointer_hash() -- which is now therefore different from the
value returned by the faster and simpler pointer_hash_fast() -- or for hashing
routines that compose various simpler routines such as zone_hash() or even
stats_hash().