node_g2_read(): need to only read 3 bytes usually to determine frame length.
Also added assertion that guarantees we have not read more bytes already than
what the frame can hold: we want to parse garbage safely.

Previously, we were reading 4 bytes to determine the frame length and when
handling garbage or mis-parsing incoming data, we could end-up being in a
setup where the computed frame length was 3 bytes but we had already read 4.
That would cause havoc later on when attempting to pmsg_read() -1 bytes!