SDBM: added concept of "loose" iterations over the database.
When the database is large, looping over all its entries not only requires
the database to be locked, but it also takes time.  This can potentially
block (i.e. freeze) an application.

A loose iteration can happen concurrently in a separate thread, without
blocking other database accesses, provided the application can cope with
some of the limitations, such as: entries can be processed more than once,
the entry is not necessarily the latest version, some entries may be missed.