VMM: fixed race condition in vmm_munmap() that could lead to assertion failures.
We were wrongly unmapping the region BEFORE updating the pmap.  Since the
munmap() system call is not performed whilst holding the lock on the pmap,
on purpose, this gave an opportunity to a concurrent thread to allocate
something in the VM space before we could mark that space as available!