omalloc: fixed chunk management when memory alignment is required.
When we have to adjust the allocation pointer for allocations, we cannot
compute the size of the chunk anymore, because it could have changed due
the alignment of the first allocation point.

Therefore, it is necessary to capture the original size of the chunk, to
determine the proper indexing chunk list where the chunk was located before
the alignment took place.