sdbm_close(): loudly warn but do not refuse to process DBs with refcnt > 1.
When sdbm_ref() is used, one should consistently use sdbm_unref() so that
the DB is properly closed when the last reference is gone.

To ease conversion from a mono-threaded to a multi-threaded use of SDBM files,
the sdbm_close() routine now simply removes the extra refefence but loudly
warns when it is invoked with a multiply-referenced database.

sdbm_close_internal(): when destroy is FALSE, it's OK if db->refcnt > 1.

Since when the "destroy" argument is FALSE, the aim is not to destroy the
object, it is perfectly fine if there are still some extra references.