crash_invoke_inspector(): also attempt to launch gdb on Windows.
Now that we have spopen(), which works on both UNIX and Windows platforms, we
can attempt to launch gdb upon a crash on Windows, which will work if they
have installed Cygwin (possible even if GTKG was not compiled by user) or run
under MinGW (unlikely, only for developers).

We still keep the proven-to-work fork() handling code on UNIX platforms, since
spopen() can allocate memory and that could create problems within a crash.