mingw_vmm_init(): accelerate initialization by reducing amount of system calls.
Increased the granularity of memory adjustements up to 4 MiB (typically is
64 KiB only at the system level), to speed-up the first pass where we
determine the available VM space we can dispose of.

Furthermore, there is no need to iterate since we can formally compute the
VM size we want to reserve, assuming the available VM space is constant
regardless of the percentage of the memory space we reserve.