Windows: stop fatal recursion when the VMM layer extends the pmap.
On Windows, we frown upon allocations with NULL hints, and try to warn with
a stack dump when they happen. However, this may require memory allocation and
when the VMM is in the process of extending its pmap, no further allocation
can occur or an assertion will fail: nothing can be recorded in the pmap since
it is full.