VMM: added new flag to signal OOM condition and be aggressive with the cache.
In OOM conditions, even if we are flagged as "crashing" currently, we strive
to allocate from the page cache as much as we can, continuing to free memory
into the cache as well to assure that the basic memory requirements up to
the point where we can safely exit or auto-restart are met.

We also make sure freed memory is redirected to the page cache to limit the
pmap updates and ensure we have sufficient pages to honour further memory
requests, for as long as possible.