signal_abort(): removed the "no return" hint, and make sure we don't s_error().
It's useless to call s_error() if we get back from raise(), since that will
create an recursion via log_abort() which will call signal_abort().

The contract must be to raise a SIGABRT, and let our caller handle the fact
that we return, instead of doing error recovery everywhere.