Added a new crash-level for deadlocks.
This allows disabling memory allocators when a deadlock happens in them,
to make sure we can continue without too much memory corruption in case
the crash code uses this deadlocked allocator.

On deadlocks, suspend the other threads and log origin quickly before handling.

In case the exception handling code does not go far enough to log an error,
the fact that we have deadlocked is registered in the crash variables so that
we can outline that fact in the crash log.

New strategy in lock "deadlock" routines: ignore when in pass-through mode.

The rationale is that we have activated a crash mode, locks are therefore
disabled and will be pass-through.  It does not matter if we already own
a spinlock, grab it again, collect information.  The old strategy to abort
on recursion was preventing useful gathering of clues: why did we deadlock
in the first place?