Removed recording of first lock point now that we always have a stack lock.
Because the stack lock is guaranteed to exist when we get our thread element,
there is no longer any need to record the first stack lock and handle cases
where a lock would not be recorded in the lock stack.

Furthermore, with a lock stack always created, there is no more need for
default lock probes: all the locks will be recorded in the thread that grabs
them, provided they are not hidden purposedly.