Added necessary plumbing to allow longjmp() or siglongjmp() in a signal handler.
Because we are now maintaining whether we are in a signal handler, we need
to trap longjmp() and siglongjmp() to be able to clear the fact that we were
in a signal handler otherwise malloc() and friends will wrongly warn that
they are used in a signal handler, where in fact we have exited it already.

Application code must use Setjmp() and Sigsetjmp() in order for the trapping
code to be exercised since we cannot wrap setjmp() and sigsetjmp() in another
function, obviously.