Keep track of "once" initializations in case we end-up recursing.
If the recursion happens "too soon" we may not have initialized a lot of
internal stuff and therefore we may not be able to get a stacktrace.

Therefore, we now manually track the routines that are initialized to be
able to determine the order that is creating a problem.  We will not get a
full stack trace though, only the name of the routines being initialized.