crash_assert_logv(): don't call crash_mode() if assertion failure was recorded.
Normally, crash_assert_logv() can only be called after crash_assert_failure(),
and the latter will already call crash_mode().  Therefore, we must not do it
or it would give the false impression that we are recursively crashing.

However, as a safety precaution, call crash_mode() and loudly complain if it
appears crash_assert_failure() was not already called before.