Use even HTTP outgoing bandwidth as the last measure when stalling.
Swapped the last two measures we had: we now first disable the stealing
of unused HTTP outgoing bandwidth first (to avoid other schedulers using
that to make our uploads stall) and if stalling still occurs, then we
switch to the uniform outgoing bandwidth scheduling.

Indeed, even HTTP bandwidth is severely penalizing everyone trying to
upload, which is a shame when only a handful of sources are stalling
because they cannot cope with the traffic on their end.