[GTK2] Show proper average data speed on completed uploads.
The core side is now updating the "start_date" field in uploads once the
HTTP status headers have been sent.  This allows us to more aggressively
compute the actual data transfer rate, as long as the transfer time, and
oppose that to the total time to complete the uploaded chunk: the
diffrence between the two is the time we had to wait for the request
plus the time we spent sending back the HTTP headers.