bg_task_finished(): now calls safer bg_task_dead().
The new bg_task_dead() routine inserts the task in the dead task list,
but makes sure it is removed from the sleep queue if needed.  Indeed,
when we call bg_task_finished() and there is still a reference on the
task, we do not mark it as dead yet but put it into sleep.

There was an awful bug when we did not do that: any task terminating
with a reference count greater than one was corrupting the sleep queue
as the item was moved to the dead task list and later freed without the
sleep queue list being told about that.  Hence any reference to the dead
task from the sleep queue was creating a ticking bomb!

This was hapenning when inserting may files to the download queue when
partial file sharing was on: the insertion of the file triggered an
update of the QRT, which creates background tasks.  Ultimately, it led
to some memory corruption and inconsistency.