Allow TTH computation for seeded files.
When we have completed downloading a file that can be shared, update its
TTH information by recomputing the TTH from scratch (using our optimal
depth) in case they also want to upload the THEX information for that
file.

Also, when file is completed, we can free-up the downloading parts of
its fileinfo to reclaim memory.