Was unduly offsetting stacktrace by one extra level.
Sometimes the compiler optimizes the stack too much and we cannot see
the top routine where the assertion is failing.  Therefore, better not
offset the stack too much: if we see our internal assertion logging
routines traced, this is not dramatic; but stripping the failing routine
from the stack is.