[GTK2] Catch column resize events to persist widths quickly.
If the program crashed and did not exit cleanly, all column resizing was
lost!  By catching the resize events, we are able to alter the GUI
property immediately, which will trigger a property file update a few
seconds later.

This also simplifies shutdown code: no longer need to persist the widths
as they are now continuously persisted.