recursive_scan_step_build_partial_table(): removed assertion.
It is now possible to have a concurrent action on the GUI (removing
a file from the fileinfo list) that will clear the sf->fi pointer of
the shared file.

Therefore, we cannot assert that the entries we are processing in that
routine satisfy the shared_file_is_partial() predicate.  Instead, we
silently skip over them.