wmove(): use tmalloc_smart() with long-term memory management.
When zmove() fails to find a new spot, it can either mean that the zone
is not garbage-collected or that zalloc() was unable to find a better
spot.  However, since memory cached in magazines is not considered, it
pays to scan for a better place when running under a long-term memory
mangement scheme.