OWLIST_remove_after(): behave as OWLIST_shift() if NULL sibling.
This is for "continuity" purposes and avoid having to special-case when
we can remove items without a priori knowledge of whether they are at
the head or within the list (we just know their previous item, which
means NULL if they are at the head).

Also fixed typo in comment.