http_data_ind(): could not process chunked data!
It was a mistake to use pmsg_start() as opposed to pmsg_read_base() to
access the data returned by the RX layer.

Indeed, the dechunking code uses copy avoidance and is not necessarily
returning data in whole message blocks: it can refer to embedded data in
message blocks by manipulating the read and write pointers of the
message block to only refer to a sub-part of another physical data block.

By using pmsg_start(), we were defeating that optimization and also
assuming that all data returned by the RX stack would be in dedicated
message blocks, which is wrong.

This bug prevented correct parsing of data returned by HTTP/1.1 and
using a "chunked" transfer-encoding.