Added logfilter support.
The logfilter is configured using a rule file installed in
~/.gtk-gnutella/logfilter and which defines filtering actions to be
applied on every log message that it traps.

The filter is given the log message, the routine where it is emitted,
the source location where the log is emitted, and the log level.

Based on that information, it can extract logs to copy them to another
file based on regular expression matching, discard logs, color them,
highlight parts of the messages for easier visual inspection, add
additional tags to messages for later "grep-ing".

It can also add or suppress routine name in log messages, add or
suppress the source file location, without requiring to recompile the
application to tune this logging.

The logfilter rule file is monitored and reloaded when it changes within
the next 30 seconds at most.