Enlarged message[] buffer, as cstr_bcpy() was complaining.
Was getting this in the logs since the introduction of cstr_bcpy():

cstr_bcpy(): truncated output after 128 bytes, needed 145
    node_get_status() "core/nodes.c:12403"
	...

Added 64 more bytes (50%) to the buffer, until new warnings pop-up.

Note that the field being copied from the node structure is 256-byte
long, but it is probably too large and this is what we are attempting
to find out, empirically :-).