Semi-reliable UDP: enhanced lingering mode.
When entering lingering mode, we are now flushing all the
un-acknowledged fragments if we got at least half of the fragments
acknowledged already.

The initial alpha (parallelism value) is now initialized to 3 and is
decreased linearly, not exponentially.

The resend timeout for the first fragment (when we are sending in
cautious mode) is now growing linearly and not exponentially for
the first 3 attempts.