prop_crash_dump(): avoid subsequent errors by checking argument!
If the gnet_prop_shutdown() routine has already been called, a further
crash will invoke gnet_prop_crash_dump() nonetheless, at which point we
will be getting a NULL property set pointer!

Hence avoid a SIGSEGV by doing nothing in that case: we are in an
exception path, we need to be absolutely safe.