UPNP: fixed XML logging.
Added xml_indent_buf() for proper XML logging from UPNP: the HTTP data
buffer we get is not necessarily NUL-terminated!.

xml_indent(): changed signature to optionally return length of string.

We now quickly assess whether the input is correct by checking it ends-up
with a '>' character, at least.

Complain loudly if the input does not look like it is properly terminated.
In that case, simply duplicate the input buffer in that case, as a string.

Added trailing assertion that ensures our lexical formatting is
properly ending with a '>', since we validated the input had one.

The new parameter of xml_indent() is filled with the length of the
generated string as a convenience, if needed.  This is useful when the
string is to be printed since then we do not have to call strlen() on it.