Factorized prefix-matching from sorted array.
The new routine bsearch_prefix() will lookup a key in a sorted array
using prefix-matching, which can result in multiple matching being
reported.

The new routine bsearch_matching() can be used after bsearch_prefix()
when multiple matching is reported, to extract all the possibly matching
entries for the key.

We are now using this new code for command-line option parsing.