Windows: if not a "registered product", leave logs in current dir.
When they do not call product_init() before progstart(), then it is
certainly not a "registered product" for which we want to keep logs in
the Windows "Documents" folder.  Rather, leave them in the current
directory where the program is launched.

The progstart() call will detect that no product_init() was done and
will fake a program name forcefully, based on the executable basename,
and flag that forcing, which can be tested by product_has_forced_name().