ut_frag_pmsg_free(): be more robust, avoid internal assertion.
It was actually proven by actual execution that the assertion in that
routine failed during operations.  Since the original code was written,
the TX layer behaviour was altered to add flushing before the message
expires, and this probably changed an initial expectation.

This is another case of my being too assertive on intial design
constraints for no other reason than to validate my initial train of
thoughts!  Then comes refactoring years later...