Enhanced TRACK_MALLOC debugging power with MALLOC_CATCH_MALLOC.
By default, we are now catching malloc() library calls with TRACK_MALLOC
to be able to spot allocations done by library routines.

When facing freeing of an unknown block, we are also probing zalloc()
and xmalloc() to check whether we're not accidentally attempting a
free() on a block allocated by these other layers.