tmalloc_thread_exiting(): was not clearing local allocator.
This means we were not re-entering tmalloc_thread_create() when
allocating from the thread-magazine allocator on an "old" thread (one
that was created and which already exited), meaning we were not
re-registering the thread_atexit() callback on that new thread.

We were also not releasing the objects in the thread magazines back to
the depot (for other threads to possibly use).

Since we're not creating a lot of new threads dynamically, threads doing
heavy memory allocation, the impact was low.  But still, not clean.