evq_trampoline(): cleanup event if cannot signal thread.
When we cannot signal the remote thread about the event, we need to
properly clean it up to avoid memory leaks.

This means we need to keep the local queue locked while calling
thread_kill() to be able to properly remove the event.

Dropping code was also factored into a common path to avoid duplication.