Let emulated alloca() cleanup when thread exits.
Our emulated alloca() routine allocates a chunk and records it as a
thread-private variable.  However, it did not register a cleaning
callback, resulting in a memory leak each time a thread using our
emulated alloca() would exit.

In the process of fixing that bug, we had to export ck_destroy() because
the structure used to manage the allocations is held within the chunk so
we cannot use ck_destroy_null() as the nullification of the chunk
pointer would otherwise write in freed memory!