Use gradual bandwidth penalty when uploads are stalling.
When uploads start to early stall, we immediately decrease the amount
of bandwidth they get for the current range we are serving.  As they
unstall, we re-increase the allocated bandwidth but less than what we
decreased it by.

Also, when starting a new range, we immediately start with a penalty if
we know the remote address has been stalling recently.

As an added benefit, we no longer reduce the amount of slots per hosts to
1 when there are stalling uploads known to that address: we only consider
reducing it down to 1 when there were obvious stalling very recently.