Forcefully reduce request chunk if b/w is limited.
When a host requests too large a chunk, given what we know about the
upload bandwidth we're going to dedicate to the request, limit the size
of the data we'll be serving to aim for a reasonable transfer time (20
seconds), given that we'll at least serve 16 KiB, no matter how long it
will take, to avoid fragmenting their requests too much.

The rationale is to re-evaluate the available b/w frequently, so that we
can quickly adjust the transfer speed up or down when b/w varies.