Emit a Retry-After header on 416 replies.
Whenever we cannot satisfy a request because we're missing parts of the
requested chunk, a 416 error is returned, letting the other party know
about the chunks we can serve.

However, if the file we have is partial and has not been updated for
some time, it is a waste to let the other party "frequently" request
chunks that, in all likelyhood, are not any more available than they
were before.

As such, when we detect little activity on a partial file, start to emit
a Retry-After header to let them space out their requests.

The download code was updated to learn to ignore Retry-After delays when
it gets a 416 error and it wants to immediately retry a request, picking
one of the available chunks advertised by the remote host as part of the
416 reply.

However, this means older gtk-gnutella versions will no longer be able
to immediately request an available chunk as part of the follow-up
request.  It's deemed OK because the current version 1.1.15 is about to
become old in about a month, and 1.1.16 will be released by then and
fully compatible with the new logic.