Avoid banning hosts issuing legitimate upload requests.
When remote host has many queued downloads targetting us and we were
out for a while, it is possible that it would send many requests in a
short period of time.

When these requests end-up actively queued or are given an upload slot,
there is no harm so we can mark the connection legitimate to not count
towards the frequent connection limits we impose.