Make Content-Range and Content-Length prioritary headers.
When sending back headers in the HTTP status for a downloading host, it
is critical to have at least a Content-Range header (if content is
partial) or at least a Content-Length one.

We now make generation of these headers a priority by flagging their
callback as such (a new feature of http_send_status() introduced a few
commits ago!).