ban_force(): read incoming data before shutdown().
To be able to account bandwidth to the proper scheduler, the signature
changed to supply a banning category, from which we can infer the right
bandwidth scheduler.

Note that this does not cause extra bandwidth to be used, we're simply
reading whatever we have received (normally this is a small amount, and
the incoming bandwidth has already been consumed anyway).