http_send_status(): reworked to limit emergency aborts.
We now retry twice if we have prioritary callbacks, and only the second
time do we limit to processing those.

Also made the upload_http_content_urn_add() callback prioritary when
sending back a 416 header, reworking its logic of generation to give
priority to available ranges when sending back a 416!