Optimize ripening table when no value is kept.
A ripening table can be configured to hold no values, in which case it
will use less memory by not storing any value pointer for each of its
entries.

In that case, ripening_insert_key() must be used, and naturally calls
like ripening_update(), ripening_lookup() or ripening_insert() are
forbidden.

All callbacks that specify a value have NULL supplied when a table is
configured for keys only.