str_ichar() and str_istr() can now append.
The "outside bound" limit now start one character after the actual end
of the string, which allows one to append a char or a string at the
end of the string using str_ichar() or str_istr().

In particular, this makes it easier to use these routines to insert
something at position 0, even if the string is empty.