Show command to clean stale lockfile.
When launching gtk-gnutella and there is a stale lockfile present,
listing some PID, and it happens that there is already another process
running under this PID, we abort, complaining that another gtk-gnutella
is running.

However, if the lockfile is really stale, and the PID listed in there
has nothing to do with gtk-gnutella at the moment, the user will need to
remove the stale lockfile.

Hence we show them the command to run, listing the full lockfile path so
that it is easy to overcome this little inconvenience!