REMI: introducing byte-code and matching interpreter.
REMI is a Matching Interpreter used to perform regular expression
matching.

Compared to the original C matching engine which was really performing
a dynamic interpretation of the compiled regular expression tree, this
new engine actually interprets low-level byte code generated from the
expression tree..

There are two main advantages to REMI:

- The byte code is a more compact representation of the regex, compared
  to the whole expression tree.  Hence it exhibits better local memory
  accesses at runtime, allowing the CPU cache to be more efficient. And
  since the interpreter code is held at a tight location, we can also
  benefit from the CPU instruction cache.

- The amount of stack space used at runtime is much lower than the one
  required with the C engine.

Overall, the matching performance of REMI can be superior to that of the
C engine by a factor of 2, and allows backtracking matches where the C
engine would blow up its stack space!

This performance also comes from other low-level optimizations we can
perform when generating the byte code, which cannot be performed at the
level of the expression tree.