Add public VMM validate/invalidate routines.
We are now stricter when page invalidation is turned on: we invalidate
pages that we put back into the magazines, and we also invalidate the
memory allocated from the emergency region, requiring that the routines
performing validation and invalidation become public.

However, these routines still depend on the internal compilation options
within lib/vmm.c to activate the memory protection scheme.  So just
because they are exported does not mean they actually do something.