Fix subtle recursive resizing bug.
When using hashtables for memory management, it was possible to trigger
a resizing and, during the course of operations, have further memory
used, requiring tracking of new items in the same hashtable we were
in the process of resizing.

That led to double VMM freeing of the region used to store the data.