For time sync over UDP, write replying time later!
Although on the sending side we were already paying attention to writing
the timestamp at the lowest possible time (the UDP TX scheduler), we were
not using the same logic on the replying side.

Hence we were only updating the reply time from the message queue, but
there could be further delaying of the message by the UDP TX scheduler
after the message was popped off the message queue.