ev_link(): insert events in the proper bucket!
Legacy code was attempting to be "smart" by checking whether the
triggering time fell before the current time, in which case it was
forcing the current (processed) bucket as the target.

This led to events possibly being linked to buckets to which they did
not really belong (based on their triggering time) and was the root of
some inconsistencies, later leading to assertion failures.

Hence now the rules are simple:

- Events are always inserted in their natural bucket.
- We ensure that cq_resched() is not given a negative delay.

These will let cq_clock() process all the relevant events, even those
scheduled at the current time.