Fix thread_lock_dump() logic.
We do not want to print locking details when the locks have been
disabled.  We could potentially see invalid information, leading to
confusion.

Also it is an error to attempt to dump the held locks when suspending a
thread holding locks: if we suspend it, it needs to stop executing as
soon as possible to avoid cascading failures.