entropy_harvest_many(): accelerate processing.
Since the pool of random data collected is given to AJE where it will be
hashed, and we will then extract AJE randomness to feed the generators,
it is not critical to have highly random data fed to that pool.

Therefore, only use simple CRC32 computations, instead of hashing everything
through SHA1, to limit CPU usage.