Our regex now uses bytecode by default.
The C regex tree is freed by default, unless RE_F_KEEP_TREE is supplied
at compilation time.  This prevents the C engine from being used. If one
tries to force the C regex matching engine with RE_X_USE_C, an error
will be returned when the regex was not compiled with RE_F_KEEP_TREE.

Since the memory footprint of the bytecode representation is much smaller
than the one for the regex tree, and given the default matching is now
done with the bytecode interpreter, it does not make sense to keep the
tree around unless explicitly requested (for debugging or inspection,
mostly via the re-test program).

This forces the storage of the min and max matching length in the regex
structure though, since the tree is no longer kept around, to be able to
supply re_match_length_min() and re_match_length_max().