hashlist: avoid race condition for moveto operations.
The hash_list_moveto_head() and hash_list_moveto_tail() routines now
return the item moved and perform the move only if the item was found.

This avoid a possible race condition with a multi-threaded hashlist
whereby the client code would first do a hash_list_lookup() and then
call one of the moving routines on the found items: without first
locking the whole hashlist object and then releasing it, the item could
be gone before the move occurs...

This is also more efficient than doing the lookup then the move, since
we have only one hashtable lookup to perform.