str_slice(): change semantics of `to'.
This is now one character past the end of the string.

This allows str_slice() to be given the same offsets for `from' and `to'
to generate an empty string.  It is also consistent with the way ranges
are captured usually.

Also factorized code to extract the substring in str_slice() and str_substr(),
which are two different interfaces to perform a substring extraction.