hash_list_moveto_{head,tail}: suppress precondition.
New usage pattern introduced by previous commit means that we can call
hash_list_moveto_head(), for instance, without any item in the hash list.

Hence the assertion is no longer a routine pre-condition but becomes an
internal check, which is already done in elist_moveto_head() in our
example here as a pre-condition.

Therefore, the precondition is no longer required and is harmful since
it prevents the new usage pattern.