thread_lock_released(): give more context on problems.
When we are facing an out-of-order lock release during the normal course
of operations, we dump a stack trace via s_error().  Otherwise, we just
have the indication of an out-of-order release, but without much
information about the location of the problem, nor whether locks have
been disabled or not already.