thread_element_unique_thread(): loudly warn on resets.
Forgot to clear the TID in thread_element_mark_reusable_locked().

Now that this is fixed, it makes sense to loudly warn if we ever have to
reset the TID of a thread element.

Also added debugging assertions, protected by the optional THREAD_DEBUG
define, to make sure the TID field is always consistent in the
thread_element object.