Several improvements to hashing data structures.
It is now possible to set a minimum expected item count, so that the
table is properly sized, and does not drop below the minimum as we
remove items (pending possible further insertions).

One can also mark the hashed structures as read-only and then turn them
back to read-write mode.  Setting the table as read-only prevents
resizing but allows a table to be shared among threads without locking.
Or to make sure that we do not accidentally update a loaded table.

A new xxx_memsize() call returns the memory footprint of the table,
including the object structure plus all the internal data structures,
but excluding memory used by the objects stored inside.

Finally, there is a xxx_insert_if_missing() call that will only
perform the insertion when the key is missing in the table.  Under some
circumstances, this allows one to avoid checking whether a key is
present to then perform the insertion, avoiding an extra lookup in the
table.